edge_node/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pydantic>=2.6.4
requests>=2.31.0
numpy>=1.26.4
redis>=5.0.0